make the demo page more prominent in README
I never noticed there was a demo until I saw the release notes for v1.4.0 which were about the demo improvements.
I also suggest adding it as the repo website (About ⚙ > Website.)

Thanks for making this plugin! I really believe it is the way forward for proper RTL support in tailwindcss and with the introduction of JIT I hope tailwindlabs add all these utilities to core based on your plugin.